Refactor reconcilers
The first commit renames the current package from pkg/collector/reconcile to pkg/collector/reconcilers

Fixes #384
