+++ controllers/opentelemetrycollector_controller_test.go
@@ -36,7 +36,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/api/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/controllers"
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
-	"github.com/open-telemetry/opentelemetry-operator/pkg/collector/reconcile"
+	"github.com/open-telemetry/opentelemetry-operator/pkg/collector/reconcilers"
 )
 
 var logger = logf.Log.WithName("unit-tests")
@@ -221,14 +221,14 @@ func TestContinueOnRecoverableFailure(t *testing.T) {
 		Tasks: []controllers.Task{
 			{
 				Name: "should-fail",
-				Do: func(context.Context, reconcile.Params) error {
+				Do: func(context.Context, reconcilers.Params) error {
 					return errors.New("should fail!")
 				},
 				BailOnError: false,
 			},
 			{
 				Name: "should-be-called",
-				Do: func(context.Context, reconcile.Params) error {
+				Do: func(context.Context, reconcilers.Params) error {
 					taskCalled = true
 					return nil
 				},
@@ -237,7 +237,7 @@ func TestContinueOnRecoverableFailure(t *testing.T) {
 	})
 
 	// test
-	err := reconciler.RunTasks(context.Background(), reconcile.Params{})
+	err := reconciler.RunTasks(context.Background(), reconcilers.Params{})
 
 	// verify
 	assert.NoError(t, err)
@@ -258,7 +258,7 @@ func TestBreakOnUnrecoverableError(t *testing.T) {
 		Tasks: []controllers.Task{
 			{
 				Name: "should-fail",
-				Do: func(context.Context, reconcile.Params) error {
+				Do: func(context.Context, reconcilers.Params) error {
 					taskCalled = true
 					return expectedErr
 				},
@@ -266,7 +266,7 @@ func TestBreakOnUnrecoverableError(t *testing.T) {
 			},
 			{
 				Name: "should-not-be-called",
-				Do: func(context.Context, reconcile.Params) error {
+				Do: func(context.Context, reconcilers.Params) error {
 					assert.Fail(t, "should not have been called")
 					return nil
 				},
@@ -308,7 +308,7 @@ func TestSkipWhenInstanceDoesNotExist(t *testing.T) {
 		Tasks: []controllers.Task{
 			{
 				Name: "should-not-be-called",
-				Do: func(context.Context, reconcile.Params) error {
+				Do: func(context.Context, reconcilers.Params) error {
 					assert.Fail(t, "should not have been called")
 					return nil
 				},
