+++ pkg/collector/reconcilers/daemonset_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package reconcile
+package reconcilers
 
 import (
 	"context"
@@ -27,44 +27,44 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
 )
 
-func TestExpectedDeployments(t *testing.T) {
+func TestExpectedDaemonsets(t *testing.T) {
 	param := params()
-	expectedDeploy := collector.Deployment(param.Config, logger, param.Instance)
+	expectedDs := collector.DaemonSet(param.Config, logger, param.Instance)
 
-	t.Run("should create deployment", func(t *testing.T) {
-		err := expectedDeployments(context.Background(), param, []v1.Deployment{expectedDeploy})
+	t.Run("should create Daemonset", func(t *testing.T) {
+		err := expectedDaemonSets(context.Background(), param, []v1.DaemonSet{expectedDs})
 		assert.NoError(t, err)
 
-		exists, err := populateObjectIfExists(t, &v1.Deployment{}, types.NamespacedName{Namespace: "default", Name: "test-collector"})
+		exists, err := populateObjectIfExists(t, &v1.DaemonSet{}, types.NamespacedName{Namespace: "default", Name: "test-collector"})
 
 		assert.NoError(t, err)
 		assert.True(t, exists)
 
 	})
-	t.Run("should update deployment", func(t *testing.T) {
-		createObjectIfNotExists(t, "test-collector", &expectedDeploy)
-		err := expectedDeployments(context.Background(), param, []v1.Deployment{expectedDeploy})
+	t.Run("should update Daemonset", func(t *testing.T) {
+		createObjectIfNotExists(t, "test-collector", &expectedDs)
+		err := expectedDaemonSets(context.Background(), param, []v1.DaemonSet{expectedDs})
 		assert.NoError(t, err)
 
-		actual := v1.Deployment{}
+		actual := v1.DaemonSet{}
 		exists, err := populateObjectIfExists(t, &actual, types.NamespacedName{Namespace: "default", Name: "test-collector"})
 
 		assert.NoError(t, err)
 		assert.True(t, exists)
 		assert.Equal(t, instanceUID, actual.OwnerReferences[0].UID)
-		assert.Equal(t, int32(2), *actual.Spec.Replicas)
 	})
 
-	t.Run("should delete deployment", func(t *testing.T) {
+	t.Run("should delete daemonset", func(t *testing.T) {
+
 		labels := map[string]string{
 			"app.kubernetes.io/instance":   "default.test",
 			"app.kubernetes.io/managed-by": "opentelemetry-operator",
 		}
-		deploy := v1.Deployment{}
-		deploy.Name = "dummy"
-		deploy.Namespace = "default"
-		deploy.Labels = labels
-		deploy.Spec = v1.DeploymentSpec{
+		ds := v1.DaemonSet{}
+		ds.Name = "dummy"
+		ds.Namespace = "default"
+		ds.Labels = labels
+		ds.Spec = v1.DaemonSetSpec{
 			Selector: &metav1.LabelSelector{
 				MatchLabels: labels,
 			},
@@ -80,27 +80,29 @@ func TestExpectedDeployments(t *testing.T) {
 				},
 			},
 		}
-		createObjectIfNotExists(t, "dummy", &deploy)
 
-		err := deleteDeployments(context.Background(), param, []v1.Deployment{expectedDeploy})
+		createObjectIfNotExists(t, "dummy", &ds)
+
+		err := deleteDaemonSets(context.Background(), param, []v1.DaemonSet{expectedDs})
 		assert.NoError(t, err)
 
-		actual := v1.Deployment{}
+		actual := v1.DaemonSet{}
 		exists, _ := populateObjectIfExists(t, &actual, types.NamespacedName{Namespace: "default", Name: "dummy"})
 
 		assert.False(t, exists)
 
 	})
 
-	t.Run("should not delete deployment", func(t *testing.T) {
+	t.Run("should not delete daemonset", func(t *testing.T) {
+
 		labels := map[string]string{
-			"app.kubernetes.io/instance":   "default.test",
 			"app.kubernetes.io/managed-by": "helm-opentelemetry-operator",
 		}
-		deploy := v1.Deployment{}
-		deploy.Name = "dummy"
-		deploy.Namespace = "default"
-		deploy.Spec = v1.DeploymentSpec{
+		ds := v1.DaemonSet{}
+		ds.Name = "dummy"
+		ds.Namespace = "default"
+		ds.Labels = labels
+		ds.Spec = v1.DaemonSetSpec{
 			Selector: &metav1.LabelSelector{
 				MatchLabels: labels,
 			},
@@ -116,12 +118,13 @@ func TestExpectedDeployments(t *testing.T) {
 				},
 			},
 		}
-		createObjectIfNotExists(t, "dummy", &deploy)
 
-		err := deleteDeployments(context.Background(), param, []v1.Deployment{expectedDeploy})
+		createObjectIfNotExists(t, "dummy", &ds)
+
+		err := deleteDaemonSets(context.Background(), param, []v1.DaemonSet{expectedDs})
 		assert.NoError(t, err)
 
-		actual := v1.Deployment{}
+		actual := v1.DaemonSet{}
 		exists, _ := populateObjectIfExists(t, &actual, types.NamespacedName{Namespace: "default", Name: "dummy"})
 
 		assert.True(t, exists)
