+++ pkg/collector/reconcilers/deployment.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package reconcile
+package reconcilers
 
 import (
 	"context"
@@ -27,29 +27,29 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
 )
 
-// +kubebuilder:rbac:groups="apps",resources=daemonsets,verbs=get;list;watch;create;update;patch;delete
+// +kubebuilder:rbac:groups="apps",resources=deployments,verbs=get;list;watch;create;update;patch;delete
 
-// DaemonSets reconciles the daemon set(s) required for the instance in the current context.
-func DaemonSets(ctx context.Context, params Params) error {
-	desired := []appsv1.DaemonSet{}
-	if params.Instance.Spec.Mode == "daemonset" {
-		desired = append(desired, collector.DaemonSet(params.Config, params.Log, params.Instance))
+// Deployments reconciles the deployment(s) required for the instance in the current context.
+func Deployments(ctx context.Context, params Params) error {
+	desired := []appsv1.Deployment{}
+	if params.Instance.Spec.Mode == "deployment" {
+		desired = append(desired, collector.Deployment(params.Config, params.Log, params.Instance))
 	}
 
 	// first, handle the create/update parts
-	if err := expectedDaemonSets(ctx, params, desired); err != nil {
-		return fmt.Errorf("failed to reconcile the expected daemon sets: %v", err)
+	if err := expectedDeployments(ctx, params, desired); err != nil {
+		return fmt.Errorf("failed to reconcile the expected deployments: %v", err)
 	}
 
 	// then, delete the extra objects
-	if err := deleteDaemonSets(ctx, params, desired); err != nil {
-		return fmt.Errorf("failed to reconcile the daemon sets to be deleted: %v", err)
+	if err := deleteDeployments(ctx, params, desired); err != nil {
+		return fmt.Errorf("failed to reconcile the deployments to be deleted: %v", err)
 	}
 
 	return nil
 }
 
-func expectedDaemonSets(ctx context.Context, params Params, expected []appsv1.DaemonSet) error {
+func expectedDeployments(ctx context.Context, params Params, expected []appsv1.Deployment) error {
 	for _, obj := range expected {
 		desired := obj
 
@@ -57,14 +57,14 @@ func expectedDaemonSets(ctx context.Context, params Params, expected []appsv1.Da
 			return fmt.Errorf("failed to set controller reference: %w", err)
 		}
 
-		existing := &appsv1.DaemonSet{}
+		existing := &appsv1.Deployment{}
 		nns := types.NamespacedName{Namespace: desired.Namespace, Name: desired.Name}
 		err := params.Client.Get(ctx, nns, existing)
 		if err != nil && k8serrors.IsNotFound(err) {
 			if err := params.Client.Create(ctx, &desired); err != nil {
 				return fmt.Errorf("failed to create: %w", err)
 			}
-			params.Log.V(2).Info("created", "daemonset.name", desired.Name, "daemonset.namespace", desired.Namespace)
+			params.Log.V(2).Info("created", "deployment.name", desired.Name, "deployment.namespace", desired.Namespace)
 			continue
 		} else if err != nil {
 			return fmt.Errorf("failed to get: %w", err)
@@ -90,17 +90,18 @@ func expectedDaemonSets(ctx context.Context, params Params, expected []appsv1.Da
 		}
 
 		patch := client.MergeFrom(existing)
+
 		if err := params.Client.Patch(ctx, updated, patch); err != nil {
 			return fmt.Errorf("failed to apply changes: %w", err)
 		}
 
-		params.Log.V(2).Info("applied", "daemonset.name", desired.Name, "daemonset.namespace", desired.Namespace)
+		params.Log.V(2).Info("applied", "deployment.name", desired.Name, "deployment.namespace", desired.Namespace)
 	}
 
 	return nil
 }
 
-func deleteDaemonSets(ctx context.Context, params Params, expected []appsv1.DaemonSet) error {
+func deleteDeployments(ctx context.Context, params Params, expected []appsv1.Deployment) error {
 	opts := []client.ListOption{
 		client.InNamespace(params.Instance.Namespace),
 		client.MatchingLabels(map[string]string{
@@ -108,7 +109,7 @@ func deleteDaemonSets(ctx context.Context, params Params, expected []appsv1.Daem
 			"app.kubernetes.io/managed-by": "opentelemetry-operator",
 		}),
 	}
-	list := &appsv1.DaemonSetList{}
+	list := &appsv1.DeploymentList{}
 	if err := params.Client.List(ctx, list, opts...); err != nil {
 		return fmt.Errorf("failed to list: %w", err)
 	}
@@ -126,7 +127,7 @@ func deleteDaemonSets(ctx context.Context, params Params, expected []appsv1.Daem
 			if err := params.Client.Delete(ctx, &existing); err != nil {
 				return fmt.Errorf("failed to delete: %w", err)
 			}
-			params.Log.V(2).Info("deleted", "daemonset.name", existing.Name, "daemonset.namespace", existing.Namespace)
+			params.Log.V(2).Info("deleted", "deployment.name", existing.Name, "deployment.namespace", existing.Namespace)
 		}
 	}
