+++ pkg/collector/reconcilers/opentelemetry.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // Package reconcile contains reconciliation logic for OpenTelemetry Collector components.
-package reconcile
+package reconcilers
 
 import (
 	"context"
